requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.8
django-mysql==2.2.0
djangorestframework==3.7.7
fcm-django==0.2.13
flake8==3.5.0
idna==2.6
mccabe==0.6.1
mysqlclient==1.3.12
pycodestyle==2.3.1
pyfcm==1.4.3
pyflakes==1.6.0
pytz==2017.3
requests==2.18.4
requests-toolbelt==0.8.0
urllib3==1.22
